Add Goto Conferences
Name of the youtuber: GotoConferences
What the channel is about (eg. web development, design, ...):  Talks
Which section is the channel in? (if you created a section, please specify why): Programming in general.
Why do you consider the youtuber deserves a place in this list? What does make it awesome?: The channel has over 300K subscribers, One of the largest conferences Channel with a lot of new concept  .Enriching